\name{hdv2003}
\docType{data}
\alias{hdv2003}
\title{Histoire de Vie 2003}
\description{
  Échantillon de 2000 individus et de 20 variables issu de l'enquête
  \emph{Histoire de Vie} réalisé par l'INSEE en 2003.
}
\usage{data(hdv2003)}
\format{Data frame comportant 2000 lignes et 20 colonnes}
\source{Fichiers détail de l'INSEE : \url{http://www.insee.fr/fr/themes/detail.asp?ref_id=fd-HDV03}}
\keyword{datasets}