\encoding{UTF-8}
\name{rgrs-deprecated}
\alias{copie}
\alias{copie.default}
\alias{copie.proptab}
\alias{cprop}
\alias{cramer.v}
\alias{format.proptab}
\alias{freq}
\alias{lprop}
\alias{print.proptab}
\alias{prop}
\alias{quant.cut}
\alias{renomme.variable}
\alias{residus}
\alias{wtd.mean}
\alias{wtd.var}
\alias{wtd.table}


%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Fonctions remplacées par leur équivalent dans l'extension questionr }
\description{
  Ces fonctions ne sont plus maintenues et ne doivent plus être
  utilisées. Elles ont été remplacées par des fonctions équivalentes
  dans l'extension questionr.
}
\usage{ 
copie(...)
copie.default(...)
copie.proptab(...)
cprop(...)
cramer.v(...)
format.proptab(...)
freq(...)
lprop(...)
print.proptab(...)
prop(...)
quant.cut(...)
renomme.variable(...)
residus(...)
wtd.mean(...)
wtd.var(...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{...}{ arguments passés aux fonctions d'origine. }
}
\author{ Julien Barnier <julien.barnier@ens-lyon.fr>}
\keyword{univar}
