\encoding{UTF-8}
\name{rgrs-package}
\alias{rgrs-package}
\alias{rgrs}
\docType{package}
\title{
rgrs
}
\description{
Fonctions pour l'utilisation de R en sciences sociales
}
\details{
\tabular{ll}{
Package: \tab rgrs\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2008-04-07\cr
License: \tab GPL-3\cr
}

Ce paquet regroupe diverses fonctions utiles pour l'analyse de données
en sciences sociales.

}
\author{
Julien Barnier <julien.barnier@ens-lsh.fr>
}
\keyword{univar}{multivariate}{manip}
