\name{rgrs.update}
\alias{rgrs.update}
\title{ Mise à jour du paquet Rgrs }
\description{
  Vérifie si une nouvelle version de Rgrs est disponible et effectue la
  mise à jour si nécessaire.
}
\usage{
rgrs.update()
}
\details{
  Nécessite d'avoir une connexion à Internet active pour pouvoir
  contacter le serveur R-forge.
}
\author{ Julien Barnier <julien.barnier@ens-lyon.fr> }
\seealso{ \command{\link{update.packages}} }
\keyword{ utilities }

