\name{rp99}
\docType{data}
\alias{rp99}
\title{Recensement 1999 - Communes du Rhône}
\description{
  Résultats tirés du recensement de la population de 1999 pour les
  communes du Rhône.
}
\usage{data(rp99)}
\format{Data frame comportant 301  lignes et 21 colonnes}
\source{Bases de données recensement de l'INSEE : \url{http://www.insee.fr/fr/bases-de-donnees/default.asp?page=recensements.htm}}
\keyword{datasets}