\encoding{UTF-8}
\name{selectwd}
\alias{selectwd}
\title{ Sélecteur de répertoire de travail }
\description{
  Affiche ue boîte de sélection de répertoire en Tk et modifie le
  répertoire de travail selon la sélection.
}
\usage{
selectwd()
}
\value{
  La fonction renvoit le répertoire choisi sous forme de chaîne de
  caractères, et affiche la commande \command{setwd} correspondant à la
  sélection effectuée.
}
\seealso{
  \command{\link{setwd}}, \command{\link{getwd}}
}
\author{ Julien Barnier <julien.barnier@ens-lyon.fr> }
\keyword{utilities}
